refactor: move slt clients into own module (#2824)
small PR to just move all of the clients inside `crates/slt/src/test.rs`
into their own module `clients`.

No functional changes, just some refactoring to make the code a bit more
organized.
