+++ crates/slt/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod clients;
 pub mod discovery;
 pub mod hooks;
 pub mod test;
